requirements.txt
azure-functions
azure-identity
azure-storage-file-datalake
./utils
